apps/api/requirements.txt
# ===== Core API & Server =====
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
pydantic[email]
python-dotenv==1.0.1
python-multipart==0.0.9
cryptography

# ===== Database & Storage =====
sqlalchemy==2.0.36
alembic
psycopg2-binary==2.9.9
asyncpg
aiomysql
redis==5.0.8
celery==5.4.0

# ===== AI Orchestration & Graph =====
langgraph==0.2.45
langchain==0.2.16
langchain-community==0.2.16
langchain-experimental==0.0.65
langchain-google-genai==1.0.10
google-generativeai==0.7.2

# ===== Vector DB & Search =====
qdrant-client[fastembed]==1.11.0
fastembed==0.3.4
ragas==0.1.21

# ===== Document Processing (DeepDoc) =====
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
Pillow==10.4.0
nltk==3.9.1
pymupdf

# ===== Utils =====
sqlparse==0.5.1
numpy==1.26.4
httpx==0.27.2
pytest==8.3.3
prometheus-client==0.21.0
rapidfuzz==3.9.7
